+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add zeroconf/mDNS discovery
 - Support openWakeWord
 - Support multiple wake words
 - Save active wake words to preferences JSON file